Se actualiza el nombre de Queries a PokemonQueries en el QueriesManager. Se añaden las queries necesarias para pokemon, se actualiza el endpoint getPokemonsByFilter y se actualiza el enlance de lanzamiento
